commit message

+++ robit task7.cpp
@@ -1,12 +1,12 @@
 #include<stdio.h>
 int main() {
 
-	int n_pac=1, nr_pac=1, r_pac=1, n_two_r=1, ncr=1, nrma1_pac=1, nma1_pac=1;
-	int npr, nHr;
+	float n_pac=1, nr_pac=1, r_pac=1, n_two_r=1, ncr=1, nrma1_pac=1, nma1_pac=1;
+	float npr, nHr;
 
-	int n, r, i;
+	float n, r, i;
 	printf("n, r 입력 : ");
-	scanf_s("%d %d", &n, &r);
+	scanf_s("%f %f", &n, &r);
 
 	for (i = 1; i <= n; i++) //n!
 		n_pac *= i;
@@ -18,17 +18,17 @@ int main() {
 		r_pac *= i;
 
 	npr = n_pac / nr_pac;
-	printf("순열: %d\n", npr);
+	printf("순열: %f\n", npr);
 
 	for (i = 1; i <= r; i++)
 		n_two_r *= n;
 
 	
-	printf("중복 순열: %d\n", n_two_r);
+	printf("중복 순열: %f\n", n_two_r);
 
 
 	ncr = n_pac / (nr_pac * r_pac);
-	printf("조합: %d\n", ncr);
+	printf("조합: %f\n", ncr);
 
 	for (i = 1; i <= n + r - 1; i++) //(n+r-1)!
 		nrma1_pac *= i;
@@ -39,7 +39,7 @@ int main() {
 
 	nHr = nrma1_pac / (nma1_pac * r_pac);
 
-	printf("중복 조합: %d", nHr);
+	printf("중복 조합: %f", nHr);
